refactor: rename textTokenLimiter.ts to text.ts
