+++ packages/api/src/utils/index.ts
@@ -11,7 +11,7 @@ export * from './llm';
 export * from './math';
 export * from './openid';
 export * from './tempChatRetention';
-export * from './textTokenLimiter';
+export * from './text';
 export { default as Tokenizer } from './tokenizer';
 export * from './yaml';
 export * from './http';
