+++ packages/api/src/utils/text.ts
